Made tile resources an entity
This was to fix rendering issues with resources alongside entities.
We can group them and order them in one set so that we don't render
entities further right over entities further down the screen.
This will be handy later on when we implementing resource gathering.
We can treat resources the same as mobs and destroy them upon attack.